Refactor: Terminal cell의 하위 요소들을 파일로 분리
Terminal cell에 관련된 파일을 단일한 파일에서 모두 관리하고 있었다. 이에
점점 관리가 힘들어져서 이를 분리하는 작업을 수행하였다.
